refactor: git mv -f Tests/nofib/NoFib/Prelude/Permute.swift Tests/SwiftToGPUTests/Prelude/Permute.swift
